# express-api/public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"PingResponse": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Estate": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"locality": {
						"type": "string"
					},
					"url": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"locality",
					"url"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DeleteResult": {
				"description": "Result object returned by DeleteQueryBuilder execution.",
				"properties": {
					"raw": {
						"description": "Raw SQL result returned by executed query."
					},
					"affected": {
						"type": "number",
						"format": "double",
						"nullable": true,
						"description": "Number of affected rows/documents\nNot all drivers support this"
					}
				},
				"required": [
					"raw"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IEstatePayload": {
				"properties": {
					"name": {
						"type": "string"
					},
					"locality": {
						"type": "string"
					},
					"url": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"locality",
					"url"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_Estate_": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"locality": {
						"type": "string"
					},
					"url": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "express-api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/ping": {
			"get": {
				"operationId": "GetMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PingResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/estates/test": {
			"post": {
				"operationId": "Test",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Estate"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "string"
							}
						}
					}
				}
			}
		},
		"/estates": {
			"get": {
				"operationId": "GetEstates",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Estate"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Estate"
				],
				"security": [],
				"parameters": []
			},
			"delete": {
				"operationId": "DeleteEstates",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DeleteResult"
								}
							}
						}
					}
				},
				"tags": [
					"Estate"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateEstate",
				"responses": {
					"200": {
						"description": "Created estate.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Estate"
								}
							}
						}
					}
				},
				"description": "Create a single estate.",
				"tags": [
					"Estate"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "- Estate payload.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IEstatePayload",
								"description": "- Estate payload."
							}
						}
					}
				}
			}
		},
		"/estates/more": {
			"post": {
				"operationId": "CreateEstates",
				"responses": {
					"200": {
						"description": "Created estates.",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Partial_Estate_"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Create multiple estates.",
				"tags": [
					"Estate"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "- Array of estate payloads.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/IEstatePayload"
								},
								"type": "array",
								"description": "- Array of estate payloads."
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}